Updated source code for MIOpen

+++ test/gtest/cba.hpp
@@ -25,11 +25,13 @@
  *******************************************************************************/
 #pragma once
 
+#include <random>
+
+#include <gtest/gtest.h>
 #include <miopen/miopen.h>
 #include <miopen/solver_id.hpp>
-
-#include "../serialize.hpp"
-#include "../fusionHost.hpp"
+#include <serialize.hpp>
+#include <fusionHost.hpp>
 
 #include "tensor_util.hpp"
 #include "get_handle.hpp"
@@ -37,8 +39,6 @@
 
 #include "conv_test_base.hpp"
 #include "conv_tensor_gen.hpp"
-
-#include <random>
 
 template <typename T = float, typename TestCaseType = ConvTestCaseBase>
 struct ConvBiasActivInferTest : public ::testing::TestWithParam<std::tuple<miopenActivationMode_t,
@@ -106,11 +106,11 @@ protected:
     miopenActivationMode_t activ_mode;
     miopen::FusionPlanDescriptor fusePlanDesc;
     miopen::OperatorArgs params;
-    const float alpha = static_cast<float>(1.0f);
-    const float beta  = static_cast<float>(0);
-    float activ_alpha = static_cast<double>(0.25f);
-    float activ_beta  = static_cast<double>(0.75f);
-    float activ_gamma = static_cast<double>(0.5f);
+    const float alpha = 1.0f;
+    const float beta  = 0.0f;
+    float activ_alpha = 0.25f;
+    float activ_beta  = 0.75f;
+    float activ_gamma = 0.5f;
     miopenTensorLayout_t tensor_layout;
     using cfsb = ConvFwdSolverTestBase<T, T, TestCaseType>;
     Workspace wspace{};
