+++ test/gtest/cba_find2.hpp
@@ -25,13 +25,15 @@
  *******************************************************************************/
 #pragma once
 
+#include <random>
+
+#include <gtest/gtest.h>
 #include <miopen/miopen.h>
 #include <miopen/problem.hpp>
 #include <miopen/solver_id.hpp>
 #include <miopen/fusion/fusion_invoke_params.hpp>
-
-#include "../serialize.hpp"
-#include "../fusionHost.hpp"
+#include <serialize.hpp>
+#include <fusionHost.hpp>
 
 #include "tensor_util.hpp"
 #include "get_handle.hpp"
@@ -39,8 +41,6 @@
 
 #include "conv_test_base.hpp"
 #include "conv_tensor_gen.hpp"
-
-#include <random>
 
 template <typename T = float>
 struct ConvBiasActivInferFind2Test
@@ -59,11 +59,11 @@ protected:
     miopen::FusionPlanDescriptor fusePlanDesc;
     */
     miopen::OperatorArgs params;
-    const float alpha       = static_cast<float>(1.0f);
-    const float beta        = static_cast<float>(0);
-    const float activ_alpha = static_cast<double>(0.25f);
-    const float activ_beta  = static_cast<double>(0.75f);
-    const float activ_gamma = static_cast<double>(0.5f);
+    const float alpha       = 1.0f;
+    const float beta        = 0.0f;
+    const float activ_alpha = 0.25f;
+    const float activ_beta  = 0.75f;
+    const float activ_gamma = 0.5f;
     miopenTensorLayout_t tensor_layout;
 
     miopen::FusedProblem fused_problem;
