+++ test/gtest/na.hpp
@@ -26,10 +26,9 @@
 #pragma once
 
 #include <gtest/gtest.h>
-#include <miopen/miopen.h>
 #include <miopen/solver_id.hpp>
-#include "../serialize.hpp"
-#include "../fusionHost.hpp"
+#include <serialize.hpp>
+#include <fusionHost.hpp>
 
 #include "tensor_util.hpp"
 #include "get_handle.hpp"
@@ -196,10 +195,10 @@ protected:
     miopenActivationMode_t activ_mode;
     miopen::FusionPlanDescriptor fusePlanDesc;
     miopen::OperatorArgs params;
-    const float alpha       = static_cast<float>(1.0f);
-    const float beta        = static_cast<float>(0);
-    const float activ_alpha = static_cast<double>(0.5f);
-    const float activ_beta  = static_cast<double>(0.5f);
-    const float activ_gamma = static_cast<double>(0.5f);
+    const float alpha       = 1.0f;
+    const float beta        = 0.0f;
+    const float activ_alpha = 0.5f;
+    const float activ_beta  = 0.5f;
+    const float activ_gamma = 0.5f;
     double epsilon          = 1.0e-5;
 };
